pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "VietnamLeadExtractor"
version = "0.1.0"
description = "Công cụ mã nguồn mở Python tự động trích xuất leads B2B từ Google Maps."
authors = [{ name = "VietnamLeadExtractor Contributors" }]
license = { text = "MIT" }
dependencies = [
    "pandas>=2.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "click>=8.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
